moving config files to server dir
